refactor(client/game): refactor game application
