+++ apps/client/src/features/game/renderer/render-config.ts
@@ -2,21 +2,18 @@
  * Shared Pixi renderer constants for the first static grid pass.
  */
 export const RenderConfig = {
+  // Stage
   stageBackground: 0x101318,
   stagePadding: 24,
-  cellGap: 2,
-  cellStride: 100,
   minScale: 0.25,
   maxScale: 1,
-  terrainIconScale: 0.78,
 
-  // Intent arrow
-  arrowColor: 0x000000,
-  arrowStrokeColor: 0xffffff,
-  arrowStrokeWidth: 3,
-  arrowLength: 44,
-  arrowThickness: 7,
-  arrowHeadSize: 22,
+  // Grid
+  cellGap: 2,
+  cellStride: 100,
+
+  // Terrain icons
+  terrainIconScale: 0.78,
 
   // Troop count text
   troopTextFontFamily: "Arial, sans-serif",
@@ -25,4 +22,12 @@ export const RenderConfig = {
   troopTextColor: 0x000000,
   troopTextStrokeColor: 0xffffff,
   troopTextStrokeWidth: 4,
+
+  // Intent arrow
+  arrowColor: 0x000000,
+  arrowStrokeColor: 0xffffff,
+  arrowStrokeWidth: 3,
+  arrowLength: 44,
+  arrowThickness: 7,
+  arrowHeadSize: 22,
 } as const;
