+++ apps/client/src/features/game/renderer/render-grid.ts
@@ -1,0 +1,12 @@
+import type { ICoordinate, Terrain, Visibility } from "@generals-plus/engine";
+import { Grid2D } from "@generals-plus/engine";
+
+export interface RenderGridCell {
+  coordinate: ICoordinate;
+  terrain: Terrain;
+  troopCount: number | null;
+  visibility: Visibility;
+  ownerIndex: string | null;
+}
+
+export class RenderGrid extends Grid2D<RenderGridCell> {}
